+++ source/main/index.js
@@ -24,12 +24,15 @@ import {
   pubLogsFolderPath,
   RTS_FLAGS,
   stateDirectoryPath,
+  RTS_FLAGS,
+  MINIMUM_AMOUNT_OF_RAM_FOR_RTS_FLAGS,
 } from './config';
 import { setupCardanoNode } from './cardano/setup';
 import { CardanoNode } from './cardano/CardanoNode';
 import { safeExitWithCode } from './utils/safeExitWithCode';
 import { buildAppMenus } from './utils/buildAppMenus';
 import { getLocale } from './utils/getLocale';
+import { getRtsFlags, setRtsFlagsAndRestart } from './utils/rtsFlags';
 import { detectSystemLocale } from './utils/detectSystemLocale';
 import { ensureXDGDataIsSet } from './cardano/config';
 import { rebuildApplicationMenu } from './ipc/rebuild-application-menu';
@@ -177,12 +180,23 @@ const onAppReady = async () => {
   );
   saveWindowBoundsOnSizeAndPositionChange(mainWindow, requestElectronStore);
 
-  const currentRtsFlags = getRtsFlagsSettings(network) || [];
+  const getCurrentRtsFlags = () => {
+    const rtsFlagsFromStorage = getRtsFlags(network);
+    if (!rtsFlagsFromStorage) {
+      if (os.totalmem() < MINIMUM_AMOUNT_OF_RAM_FOR_RTS_FLAGS) {
+        setRtsFlagsAndRestart(environment.network, RTS_FLAGS);
+        return RTS_FLAGS;
+      }
+      return [];
+    }
+    return rtsFlagsFromStorage;
+  };
 
+  const rtsFlags = getCurrentRtsFlags();
   logger.info(
-    `Setting up Cardano Node... with flags: ${JSON.stringify(currentRtsFlags)}`
+    `Setting up Cardano Node... with flags: ${JSON.stringify(rtsFlags)}`
   );
-  cardanoNode = setupCardanoNode(launcherConfig, mainWindow, currentRtsFlags);
+  cardanoNode = setupCardanoNode(launcherConfig, mainWindow, rtsFlags);
 
   buildAppMenus(mainWindow, cardanoNode, userLocale, {
     isNavigationEnabled: false,